Added support for custom Homebrew installations.
App will now search for homebrew in “/usr/local/bin/brew”. If not found, will lock the app and request the user to install Homebrew. Also, a button was added to show the custom Homebrew installation window with instructions.
Now the Homebrew tasks run on a separate queue.
Fixed problem which caused failed operations to result in no error message.