Hack to make Cakebrew ignore output from login shell scripts.
Feel free to ignore this, but it's just one of many ideas on
how to solve this issue. ;)

Changed BPHomebrewInterface so that it wouldn't break if the
login shell wrote text to stdout before the command
(e.g. brew info) was actually run.

To accomplish this we just write a known constant string to
stdout before we run the command. Then, after the whole
thing is done running we can simply find that known string
and remove everything up to and including it, leaving us
with only the intended output of the command we ran.

This should also solve the issue reported in #59:
https://github.com/brunophilipe/Cakebrew/issues/59